# apps/be/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.BookingStatus:
    enum:
    - CONFIRMED
    - CANCELED
    - PENDING
    type: string
    x-enum-varnames:
    - BookingStatusConfirmed
    - BookingStatusCanceled
    - BookingStatusPending
...
  http.BookTicketRequest:
    properties:
      quantity:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
  http.BookingResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      quantity:
        type: integer
      status:
        $ref: '#/definitions/domain.BookingStatus'
      ticket_id:
        type: string
      ticket_name:
        type: string
      total_amount:
        type: number
      unit_price:
        type: number
      user_id:
        type: string
    type: object
...
  http.UserTicketQuotaResponse:
    properties:
      current_booked:
        type: integer
      max_booking_per_user:
        type: integer
      remaining_quota:
        type: integer
      ticket_id:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.PaginatedData:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Ticket Box Team
  description: RESTful API documentation for the Ticket Box event booking platform.
  title: Ticket Box Backend API
  version: "1.0"
paths:
...
  /tickets/{id}/book:
    post:
      consumes:
      - application/json
      description: Reserve and purchase tickets with concurrency protection against
        race conditions and overselling
      parameters:
      - description: Ticket UUID
        in: path
        name: id
        required: true
        type: string
      - description: Booking details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.BookTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Ticket booked successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.BookingResponse'
              type: object
        "400":
          description: Bad request - invalid quantity, payload, or exceeded max booking
            limit per user
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Ticket not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict - ticket sold out or insufficient stock
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Book a ticket
      tags:
      - Bookings
  /tickets/{id}/my-quota:
    get:
      description: Retrieve the booking limit, current booked count, and remaining
        quota for the authenticated user
      parameters:
      - description: Ticket UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User ticket quota retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.UserTicketQuotaResponse'
              type: object
        "400":
          description: Invalid ticket ID
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Ticket not found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user's quota for a ticket
      tags:
      - Bookings
...
  /users/me/bookings:
    get:
      description: Retrieve all bookings belonging to the currently authenticated
        user
      parameters:
      - default: 1
        description: Page number (default 1, min 1)
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page (default 10, min 1, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Bookings retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user's bookings
      tags:
      - Bookings
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by your JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"